agent/engine/docker_task_engine.go: move capability functionality to the agent/app package
