+++ agent/engine/docker_task_engine_test.go
@@ -28,10 +28,7 @@ import (
 	"github.com/aws/amazon-ecs-agent/agent/config"
 	"github.com/aws/amazon-ecs-agent/agent/credentials"
 	"github.com/aws/amazon-ecs-agent/agent/credentials/mocks"
-	"github.com/aws/amazon-ecs-agent/agent/ecs_client/model/ecs"
-	"github.com/aws/amazon-ecs-agent/agent/ecscni"
 	"github.com/aws/amazon-ecs-agent/agent/ecscni/mocks"
-	"github.com/aws/amazon-ecs-agent/agent/engine/dockerclient"
 	"github.com/aws/amazon-ecs-agent/agent/engine/dockerstate"
 	"github.com/aws/amazon-ecs-agent/agent/engine/testdata"
 	"github.com/aws/amazon-ecs-agent/agent/eventstream"
@@ -1167,218 +1164,6 @@ func TestTaskTransitionWhenStopContainerReturnsTransientErrorBeforeSucceeding(t 
 		t.Error("Should be out of events")
 	default:
 	}
-}
-
-func TestCapabilities(t *testing.T) {
-	conf := &config.Config{
-		AvailableLoggingDrivers: []dockerclient.LoggingDriver{
-			dockerclient.JSONFileDriver,
-			dockerclient.SyslogDriver,
-			dockerclient.JournaldDriver,
-			dockerclient.GelfDriver,
-			dockerclient.FluentdDriver,
-		},
-		PrivilegedDisabled:      false,
-		SELinuxCapable:          true,
-		AppArmorCapable:         true,
-		TaskENIEnabled:          true,
-		TaskCleanupWaitDuration: config.DefaultConfig().TaskCleanupWaitDuration,
-	}
-	ctrl, client, _, taskEngine, _, _ := mocks(t, conf)
-	cniClient := mock_ecscni.NewMockCNIClient(ctrl)
-	taskEngine.(*DockerTaskEngine).cniClient = cniClient
-	defer ctrl.Finish()
-
-	cniCapabilities := []string{ecscni.CapabilityAWSVPCNetworkingMode}
-	gomock.InOrder(
-		client.EXPECT().SupportedVersions().Return([]dockerclient.DockerVersion{
-			dockerclient.Version_1_17,
-			dockerclient.Version_1_18,
-		}),
-		client.EXPECT().KnownVersions().Return([]dockerclient.DockerVersion{
-			dockerclient.Version_1_17,
-			dockerclient.Version_1_18,
-			dockerclient.Version_1_19,
-		}),
-		cniClient.EXPECT().Capabilities(ecscni.ECSENIPluginName).Return(cniCapabilities, nil),
-		cniClient.EXPECT().Capabilities(ecscni.ECSBridgePluginName).Return(cniCapabilities, nil),
-		cniClient.EXPECT().Capabilities(ecscni.ECSIPAMPluginName).Return(cniCapabilities, nil),
-		cniClient.EXPECT().Version(ecscni.ECSENIPluginName).Return("v1", nil),
-	)
-
-	expectedCapabilityNames := []string{
-		"com.amazonaws.ecs.capability.privileged-container",
-		"com.amazonaws.ecs.capability.docker-remote-api.1.17",
-		"com.amazonaws.ecs.capability.docker-remote-api.1.18",
-		"com.amazonaws.ecs.capability.logging-driver.json-file",
-		"com.amazonaws.ecs.capability.logging-driver.syslog",
-		"com.amazonaws.ecs.capability.logging-driver.journald",
-		"com.amazonaws.ecs.capability.selinux",
-		"com.amazonaws.ecs.capability.apparmor",
-		attributePrefix + taskENIAttributeSuffix,
-	}
-
-	var expectedCapabilities []*ecs.Attribute
-	for _, name := range expectedCapabilityNames {
-		expectedCapabilities = append(expectedCapabilities,
-			&ecs.Attribute{Name: aws.String(name)})
-	}
-	expectedCapabilities = append(expectedCapabilities,
-		&ecs.Attribute{
-			Name:  aws.String(attributePrefix + cniPluginVersionSuffix),
-			Value: aws.String("v1"),
-		})
-
-	capabilities := taskEngine.Capabilities()
-
-	for i, expected := range expectedCapabilities {
-		assert.Equal(t, aws.StringValue(expected.Name), aws.StringValue(capabilities[i].Name))
-		assert.Equal(t, aws.StringValue(expected.Value), aws.StringValue(capabilities[i].Value))
-	}
-}
-
-func TestGetTaskENIAttributeEmptyCapabilityListFromPlugin(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	cniClient := mock_ecscni.NewMockCNIClient(ctrl)
-	cniClient.EXPECT().Capabilities(ecscni.ECSENIPluginName).Return([]string{}, nil)
-
-	taskEngine := &DockerTaskEngine{
-		cniClient: cniClient,
-		cfg: &config.Config{
-			TaskENIEnabled: true,
-		},
-	}
-	_, err := taskEngine.getTaskENIAttribute()
-	assert.Error(t, err)
-}
-
-func TestGetTaskENIAttributeErrorGettingCapabilitiesFromPlugin(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	cniClient := mock_ecscni.NewMockCNIClient(ctrl)
-	cniClient.EXPECT().Capabilities(ecscni.ECSENIPluginName).Return(nil, errors.New("error"))
-
-	taskEngine := &DockerTaskEngine{
-		cniClient: cniClient,
-		cfg: &config.Config{
-			TaskENIEnabled: true,
-		},
-	}
-	_, err := taskEngine.getTaskENIAttribute()
-	assert.Error(t, err)
-}
-
-func TestCapabilitiesECR(t *testing.T) {
-	conf := &config.Config{}
-	ctrl, client, _, taskEngine, _, _ := mocks(t, conf)
-	defer ctrl.Finish()
-
-	client.EXPECT().SupportedVersions().Return([]dockerclient.DockerVersion{
-		dockerclient.Version_1_19,
-	})
-	client.EXPECT().KnownVersions().Return(nil)
-
-	capabilities := taskEngine.Capabilities()
-
-	capMap := make(map[string]bool)
-	for _, capability := range capabilities {
-		capMap[aws.StringValue(capability.Name)] = true
-	}
-
-	_, ok := capMap["com.amazonaws.ecs.capability.ecr-auth"]
-	assert.True(t, ok, "Could not find ECR capability when expected; got capabilities %v", capabilities)
-
-}
-
-func TestCapabilitiesTaskIAMRoleForSupportedDockerVersion(t *testing.T) {
-	conf := &config.Config{
-		TaskIAMRoleEnabled: true,
-	}
-	ctrl, client, _, taskEngine, _, _ := mocks(t, conf)
-	defer ctrl.Finish()
-
-	client.EXPECT().SupportedVersions().Return([]dockerclient.DockerVersion{
-		dockerclient.Version_1_19,
-	})
-	client.EXPECT().KnownVersions().Return(nil)
-
-	capabilities := taskEngine.Capabilities()
-	capMap := make(map[string]bool)
-	for _, capability := range capabilities {
-		capMap[aws.StringValue(capability.Name)] = true
-	}
-
-	ok := capMap["com.amazonaws.ecs.capability.task-iam-role"]
-	assert.True(t, ok, "Could not find iam capability when expected; got capabilities %v", capabilities)
-}
-
-func TestCapabilitiesTaskIAMRoleForUnSupportedDockerVersion(t *testing.T) {
-	conf := &config.Config{
-		TaskIAMRoleEnabled: true,
-	}
-	ctrl, client, _, taskEngine, _, _ := mocks(t, conf)
-	defer ctrl.Finish()
-
-	client.EXPECT().SupportedVersions().Return([]dockerclient.DockerVersion{
-		dockerclient.Version_1_18,
-	})
-	client.EXPECT().KnownVersions().Return(nil)
-
-	capabilities := taskEngine.Capabilities()
-	capMap := make(map[string]bool)
-	for _, capability := range capabilities {
-		capMap[aws.StringValue(capability.Name)] = true
-	}
-
-	_, ok := capMap["com.amazonaws.ecs.capability.task-iam-role"]
-	assert.False(t, ok, "task-iam-role capability set for unsupported docker version")
-}
-
-func TestCapabilitiesTaskIAMRoleNetworkHostForSupportedDockerVersion(t *testing.T) {
-	conf := &config.Config{
-		TaskIAMRoleEnabledForNetworkHost: true,
-	}
-	ctrl, client, _, taskEngine, _, _ := mocks(t, conf)
-	defer ctrl.Finish()
-
-	client.EXPECT().SupportedVersions().Return([]dockerclient.DockerVersion{
-		dockerclient.Version_1_19,
-	})
-	client.EXPECT().KnownVersions().Return(nil)
-
-	capabilities := taskEngine.Capabilities()
-	capMap := make(map[string]bool)
-	for _, capability := range capabilities {
-		capMap[aws.StringValue(capability.Name)] = true
-	}
-
-	_, ok := capMap["com.amazonaws.ecs.capability.task-iam-role-network-host"]
-	assert.True(t, ok, "Could not find iam capability when expected; got capabilities %v", capabilities)
-}
-
-func TestCapabilitiesTaskIAMRoleNetworkHostForUnSupportedDockerVersion(t *testing.T) {
-	conf := &config.Config{
-		TaskIAMRoleEnabledForNetworkHost: true,
-	}
-	ctrl, client, _, taskEngine, _, _ := mocks(t, conf)
-	defer ctrl.Finish()
-
-	client.EXPECT().SupportedVersions().Return([]dockerclient.DockerVersion{
-		dockerclient.Version_1_18,
-	})
-	client.EXPECT().KnownVersions().Return(nil)
-
-	capabilities := taskEngine.Capabilities()
-	capMap := make(map[string]bool)
-	for _, capability := range capabilities {
-		capMap[aws.StringValue(capability.Name)] = true
-	}
-
-	_, ok := capMap["com.amazonaws.ecs.capability.task-iam-role-network-host"]
-	assert.False(t, ok, "task-iam-role capability set for unsupported docker version")
 }
 
 func TestGetTaskByArn(t *testing.T) {
